Kotlin unsigned types proper support
Instead of casting the unsigned types to signed and back
whenever they are passed across the boundary, here we add
support through the IntegerType in JNA. This is how unsigned
types are supposed to be passed across the boundary.

+++ feature_tests/kotlin/somelib/src/main/kotlin/dev/diplomattest/somelib/MyIndexer.kt
@@ -8,7 +8,7 @@ import com.sun.jna.Structure
 
 internal interface MyIndexerLib: Library {
     fun namespace_MyIndexer_destroy(handle: Pointer)
-    fun namespace_MyIndexer_get(handle: Pointer, i: Long): OptionSlice
+    fun namespace_MyIndexer_get(handle: Pointer, i: size_t): OptionSlice
 }
 
 class MyIndexer internal constructor (
@@ -31,7 +31,7 @@ class MyIndexer internal constructor (
     
     internal fun getInternal(i: ULong): String? {
         
-        val returnVal = lib.namespace_MyIndexer_get(handle, i.toLong());
+        val returnVal = lib.namespace_MyIndexer_get(handle, size_t(i));
         
         val intermediateOption = returnVal.option() ?: return null
             return PrimitiveArrayTools.getUtf8(intermediateOption)
